pyproject.toml
[tool.poetry]
name = "repl_python3_jeison"
version = "0.1.0"
description = ""
authors = ["Guala <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
youtube_dl = "^2021.6.6"
discord = "^1.7.3"
python-decouple = "^3.4"
bs4 = "^0.0.1"
PyNaCl = "^1.4.0"
praw = "^7.4.0"
requests = "^2.26.0"
opencv-python = "^4.5.3"
utils = "^1.0.1"
certifi = "^2021.5.30"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
